requirements.txt
pandas
numpy
matplotlib
seaborn
flask
python-dotenv
pymongo
certifi
pymongo[srv]
scikit-learn
dill
pyaml
mlflow

#-e .